+++ src/main/java/com/vbokhan/informationhandling/interpreter/TerminalExpressionMinus.java
@@ -4,8 +4,8 @@ package com.vbokhan.informationhandling.interpreter;
 public class TerminalExpressionMinus extends AbstractMathExpression {
     @Override
     public void interpret(Context c) {
-        int first = c.popValue();
-        int second = c.popValue();
+        double first = c.popValue();
+        double second = c.popValue();
         c.pushValue(second - first);
     }
 }
